+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Meta: Locked dependencies via `poetry.lock` file in version control
 - Meta: Add [pre-commit](https://pre-commit.com) config to enforce code styling and file formatting
 
+### Fixed
+- Fixed DeprecationWarnings for using `assertEquals` in tests
+
 ### Changed
 - Meta: Replace setuptools with poetry for packaging and dependency management